Add middleware for fetching and validating submissions
Every submission route requires a valid submission. Add a middleware
to the submissions router which fetches the submission id from the
request and fetches the corresponding submission from the API. A
404 response is returned if the submission id or submission is
not found.

Update the verify handler to use the submission set by the
middleware instead of fetching it from the API directly.