Add type for user identifiers
Add ExternalUserId type which represents the type of identifiers
for users in the system. Add ExternalUser interface which represents
users with an external id.

Add LoggedInUser type and use that to represent users with a login.
This should currently only be practitioners.

Update all references to practitioner ids to use the new
ExternalUserId type instead of strings. Fetch the practitioner
identifier using the externalId method instead of by referencing
fields directly. The field used to identify practitioners within
the system is changing to the username.